重构主界面布局以提升可拓展性
- 将侧栏、工具栏、内容区从 App.tsx 抽取为独立布局组件
- 工具栏改为弹性布局,新增控件不再挤压既有元素
- 壳层几何(侧栏宽度/标题栏高度/抽屉间距)统一为 CSS 令牌
