+++ .release-notes/extensible-ui-layout.md
@@ -1,0 +1,10 @@
+# 更灵活的主界面布局
+
+## 新增功能
+
+- 🧭 顶部工具栏改为自适应排列：窗口变窄时，搜索框、筛选器和功能按钮会自动调整占位、彼此让位，不再相互挤压或溢出。
+- ➕ 工具栏现在可以自然容纳更多功能入口，后续新增按钮或筛选项时界面排布依然整齐。
+
+## Bug 修复
+
+- 修复了窗口较窄时顶部筛选区可能被截断、与右侧按钮争夺空间的问题。
